Fixes in app after merge
